Updated game flow

+++ src/Player/ArrOfPlayer.cpp
@@ -52,7 +52,7 @@ void ArrOfPlayer::changePlayerName(int idx, string name){
     this->arr[idx].setName(name);
 }
 
-void ArrOfPlayer::setPlayerPoint(int idx, int point){
+void ArrOfPlayer::setPlayerPoint(int idx, long long point){
     this->arr[idx].setPoint(point);
 }
