+++ src/Player/ArrOfPlayer.hpp
@@ -20,7 +20,7 @@ class ArrOfPlayer {
         void reverseTurn(int);
         Player getPlayer(int);
         void changePlayerName(int, string);
-        void setPlayerPoint(int, int);
+        void setPlayerPoint(int, long long);
         void addPlayerCard(int, DeckCard&);
         void addAbilityCard(int, AbilityCard&);
         void clearCard(int i);
